+++ client/src/utils/api/index.ts
@@ -1,5 +1,5 @@
-// utils/api/index.ts
+// Rooms operations
+export { GET_ROOMS, DELETE_ROOMS_AND_TASKS } from "./rooms/index";
 
 // Room operations
-export { CREATE_ROOM } from "./rooms/mutations";
-export { GET_ROOMS } from "./rooms/queries";
+export { CREATE_ROOM, GET_ROOM, UPDATE_ROOM, DELETE_ROOM } from "./room/index";
