+++ client/src/utils/api/room/index.ts
@@ -1,0 +1,4 @@
+import { GET_ROOM } from "./queries";
+import { CREATE_ROOM, UPDATE_ROOM, DELETE_ROOM } from "./mutations";
+
+export { CREATE_ROOM, GET_ROOM, UPDATE_ROOM, DELETE_ROOM };
