+++ client/src/utils/api/rooms/index.ts
@@ -1,0 +1,4 @@
+import { GET_ROOMS } from "./queries";
+import { DELETE_ROOMS_AND_TASKS } from "./mutations";
+
+export { GET_ROOMS, DELETE_ROOMS_AND_TASKS };
