+++ client/src/utils/api/rooms/queries.ts
@@ -1,10 +1,14 @@
 import { gql } from "@apollo/client";
 
 export const GET_ROOMS = gql`
-  query GetRooms {
+  query Rooms {
     rooms {
       _id
       name
+      tasks {
+        _id
+        name
+      }
     }
   }
 `;
